many fixes

+++ src/star_formation/EAGLE/star_formation.h
@@ -582,12 +582,12 @@ INLINE static void star_formation_copy_properties(
     const int convert_part) {
 
   /* Store the current mass */
-  sp->mass = hydro_get_mass(p)/double(starform->n_stars_per_sf);
+  sp->mass = hydro_get_mass(p)/(starform->n_stars_per_sf);
 
   /* Store the current mass as the initial mass */
-  sp->mass_init = hydro_get_mass(p)/double(starform->n_stars_per_sf);
+  sp->mass_init = hydro_get_mass(p)/(starform->n_stars_per_sf);
 
-  sp->gpart->mass = hydro_get_mass(p)/double(starform->n_stars_per_sf);
+  sp->gpart->mass = hydro_get_mass(p)/(starform->n_stars_per_sf);
   
   /* Store either the birth_scale_factor or birth_time depending  */
   if (with_cosmology) {
